Add SyncsFiles entity and related DTOs and interfaces
