+++ .vscode/settings.json
@@ -31,8 +31,7 @@
   ],
   "python.sortImports.path": "isort",
   "python.linting.mypyEnabled": true,
-  "python.defaultInterpreterPath": "pipenv",
-  "python.envFile": "${workspaceFolder}/backend/.env",
+  "python.defaultInterpreterPath": "python3",
   "python.linting.enabled": true,
   "python.linting.flake8Enabled": true,
   "python.linting.pycodestyleEnabled": true,
