+++ backend/modules/sync/repository/sync_interfaces.py
@@ -2,15 +2,17 @@ from abc import ABC, abstractmethod
 from uuid import UUID
 
 from modules.sync.dto.inputs import (
+    SyncFileInput,
+    SyncFileUpdateInput,
     SyncsActiveInput,
     SyncsActiveUpdateInput,
     SyncsUserInput,
     SyncUserUpdateInput,
 )
-from modules.sync.entity.sync import SyncsActive
+from modules.sync.entity.sync import SyncsActive, SyncsFiles
 
 
-class SyncInterface(ABC):
+class SyncUserInterface(ABC):
     @abstractmethod
     def create_sync_user(
         self,
@@ -19,7 +21,7 @@ class SyncInterface(ABC):
         pass
 
     @abstractmethod
-    def get_syncs_user(self,  user_id: str,sync_user_id: int = None):
+    def get_syncs_user(self, user_id: str, sync_user_id: int = None):
         pass
 
     @abstractmethod
@@ -35,6 +37,15 @@ class SyncInterface(ABC):
         self, sync_user_id: str, state: dict, sync_user_input: SyncUserUpdateInput
     ):
         pass
+
+    @abstractmethod
+    def get_files_folder_user_sync(
+        self, sync_active_id: int, user_id: str, folder_id: int = None
+    ):
+        pass
+
+
+class SyncInterface(ABC):
 
     @abstractmethod
     def create_sync_active(
@@ -59,13 +70,27 @@ class SyncInterface(ABC):
         pass
 
     @abstractmethod
-    def get_files_folder_user_sync(self, sync_active_id: int, user_id: str, folder_id: int = None):
+    def get_details_sync_active(self, sync_active_id: int):
         pass
 
     @abstractmethod
-    def get_details_sync_active(self, sync_active_id: int):
-        pass
-    
-    @abstractmethod
     async def get_syncs_active_in_interval(self):
+        pass
+
+
+class SyncFileInterface(ABC):
+    @abstractmethod
+    def create_sync_file(self, sync_file_input: SyncFileInput) -> SyncsFiles:
+        pass
+
+    @abstractmethod
+    def get_sync_files(self, sync_active_id: int) -> list[SyncsFiles]:
+        pass
+
+    @abstractmethod
+    def update_sync_file(self, sync_file_id: int, sync_file_input: SyncFileUpdateInput):
+        pass
+
+    @abstractmethod
+    def delete_sync_file(self, sync_file_id: int):
         pass
